Silence a GCC compile-time warning.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@18242 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ modules/vapor/vpr/md/POSIX/IO/Socket/InetAddrBSD.cpp
@@ -247,7 +247,8 @@ void InetAddrBSD::setFamily(const vpr::SocketTypes::Domain family)
       default:
          {
             std::ostringstream msg_stream;
-            msg_stream << "Unknown socket family value " << family;
+            msg_stream << "Unknown socket family value "
+                       << (unsigned long) family;
             throw IllegalArgumentException(msg_stream.str(), VPR_LOCATION);
          }
          break;
